+++ config/groovy/tcp-slave-agent-port.groovy
@@ -1,0 +1,11 @@
+import hudson.model.*;
+import jenkins.model.*;
+
+
+Thread.start {
+      println "--> setting agent port for jnlp"
+      def env = System.getenv()
+      int port = env['JENKINS_SLAVE_AGENT_PORT'].toInteger()
+      Jenkins.instance.setSlaveAgentPort(port)
+      println "--> setting agent port for jnlp... done"
+}